requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==2.5.1
black==22.6.0
build==0.8.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
cytoolz==0.12.0
dka-defi-sdk @ file:///Users/juusoahlroos/Documents/dka/repos/defi-sdk
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.4.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fireblocks-defi-sdk-py @ git+https://github.com/dkacapitalmgnt/fireblocks-defi-sdk-py.git@d6844a9703e81e8edc76b9d5abb8672cd7baf786
fireblocks-sdk==1.17.0
frozenlist==1.3.0
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-core==2.3.1
google-cloud-secret-manager==2.11.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
hexbytes==0.2.2
idna==3.3
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==4.6.2
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
pep517==0.12.0
pip-tools==6.8.0
platformdirs==2.5.2
pluggy==1.0.0
proto-plus==1.20.6
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
rlp==2.0.1
rsa==4.8
six==1.16.0
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.3.0
urllib3==1.26.10
varint==1.0.2
web3==5.30.0
websockets==9.1
yarl==1.7.2
